[autocomplete] Use AutoCompleteCatalogProvider abstraction
Now different Providers could be passed, with the ones with ClientContext reachable
be able to query the catalog, and others be less powerful but instantiable also in
without a ClientContext.
